Lazily initialize ValidationResult fields
 ## Why is the change being made?
Users of ValidationResult that have a lot of ValidationResult instances
see a large amount of heap space being used to store the fields
missingFields and mismatchedFields. This also has a negative performance
impact from increased GC.

 ## What has changed to address the problem?
These are usually empty, and do not need to be initialized when a
ValidationResult instance is created. Now, we create them lazily, as we
add to the fields.

 ## How was this change tested?
Current tests pass.